+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] 2023-12-30
+### Added
+- `TimeDisplay` on right site of `NavMenu`
+- In `Button` added possibility to pass customWidth prop
+
+### Changed
+- In `LoginPage` changed default Mui Button to Custombutton
+
+### Fixed
+- Hover color of `NavMenu` Buttons on big screens
+
+### Removed
+- In `VillaAnnaCountdownPage` removed console.log
+
 ## [v15.10.0] - 2023-12-30
 ### Added
 - Added `Button.tsx` for a CustomButton